Refactor the android permission and settings checks into a set of separate classes
- Move the foreground actions in `DataCollectionPlugin` to `SensorControlForegroundDelegate`
- Move the background actios from `TripDiaryStateMachineService` to `SensorControlBackgroundChecker`
- Move the shared constants to `SensorControlConstants`

This should make it easier to add the permissions for motion activity as a next step.

Testing done:
- Added the plugin
- Tried the usual changes to turn off location settings and permissions
